requirement.txt
pycryptodomex
pycryptodome
mysql-connector-python
pandas
azure.storage.blob
